requirements.txt
streamlit==1.31.1
tensorflow==2.15.0
pillow==10.2.0
pandas==2.2.0
plotly==5.18.0
scikit-learn==1.4.0
seaborn==0.13.2
numpy==1.24.3